REFARRAY: New referenced array object
This object allows creation the arrays
with the reference count. Usefull when
there are many instances of some object
have to reference dynamically allocated array
which is common for all these instances.

In case of ELAPI the event object
keeps a referecne to the common array
of the sinks in the fail over order.

We decided that it will be a common object
not specific only to ELAPI.

All the review concerns related to this
object have been addressed in this patch.
It also has been moved to the common area.