Fix path_utils_ut segfault
In the case where the allocated buffer is not large enough to hold
the resulting absolute path, we were writing out a null terminator
outside of the buffer, instead of at its beginning.

Also fixes potential issue where split_path would not initialize
the count to zero if it returned a failure.