Initialize variables in loops
It occured to me that one of the issues that Coverity
did not like (and I could not understand what it is
complaining about) is related to intializing the variables
in the loop.
This patch adds initialization in the loops.
Also there was an initialization missing in the ini_configobj.c