Fix issue when running make distcheck
When running 'make distcheck', all files in the $srcdir are set to
read-only, to ensure that the build does not modify the source.
However, this broke our ability to test file permissions. We will
now copy the file to the builddir and do the tests there, where
we have control.