Fix permission checking unit test
The unit test was broken. The wrong function was used.
To make sure everything is correct I also added a convenience
function to print the internals if the file context object.
The unit test is fixed to use relative paths consitently.
Also added nice statements at the beggining and the end of
the unit test functions where they were missing.