The implementation of the new interface
The three files in the patch implement the functions
of the new interface. It is mostly inspired by the old
interface except that the string value is replaced with
the value object that hides all the metadata about
the value itself. The main function to get value
now allows more than one value per key.