Remove code that allows dup sections
Duplicate sections in the config tree is a bad idea.
After more evaluation I decided to not allow more than one section
with the same name to be acceptable in the resulting object.
All duplicate sections need to be processed so that only
one section is left as a result. The collision flags control that.
It is OK to remove it as this code has not been released yet.
I also spotted that unit test does not exit if the system()
call was successful but the command it invoked was not.
Fixed that. Also fixed couple spelling mistakes here and there.