Improve parser
1. Move the last comment into the config object
Last comment in the file was previously treated as a special
value in the last section in the configuration object.
That created difficulties while merging. Storing an orphan
comment as part of the config object is a better approach
2. Since the structure of the config object changed
the internal header and serialization required update.
3. Parser code was updated to:
- deal with the last comment on the top level
- make sure the right error codes are returned in
different modes
- pass right information through the layers which caused
some refactoring of the internal calls
- prevent using an already used configuration object
as an input to parser.