More interface refactoring
I also realized that error list processing should not be
bound to the file object.
This patch corrects that by moving the error_list and
corresponding count from file object to the config object.
Things updated by the patch:
1. The internal variables are moved from file obj to config obj.
2. The external header is updated to reflect the change
3. Functions are moved from file obj module to config obj module.
4. Parser code is updated because error validation was in
   the wrong place
5. Unit test is adjusted to get error list from the right object.

I had to adjust the copy function for the config object.
Copy function does not copy over the error list.