Fix processing of the white space at the end of the line
Before this patch the spaces were trimmed on every line.
That causes problems if the value was folded and spaces
were at the end of the  non last line of the multi line
value. Now we trim spaces only on the last line.
To accomplish this we moved the trimming from
parser code into the layer where the value is constructed
from individual lines.