AUTOMAKE: Do not treat warnings as errors
The EXTRA_*_DEPENDENCIES variable is useful for cases where you merely want
to augment the automake-generated _DEPENDENCIES variable rather than
replacing it. It is a syntax warning in automake < 1.11.3 and
syntax warnings are enabled by default.

Makefile.am:259: variable `EXTRA_libini_config_la_DEPENDENCIES' is
    defined but no program or library has `EXTRA_libini_config_la'
    as canonical name  (possible typo)
autoreconf: automake failed with exit status: 1

The only negative of this solution is that library will not be recompiled
with older version of automake if files listed in EXTRA_*_DEPENDENCIES
are changed. It is better to see warnings than override automatically
generated dependencies calculated by automake with syntax: *_DEPENDENCIES

Reviewed-by: Sumit Bose <sbose@redhat.com>