Function to return one of the dups
The internal search functoins we not able to deal with
the duplicates that might be scattered across a collection.
This change creates a new search function that is more robust.
The old function becomes a wrapper.
The code is adjusted in the cases where the functionality
of the more advanced functions is needed.

Signed-off-by: Dmitri Pal <dpal@redhat.com>
Signed-off-by: Lukas Slebodnik <lslebodn@redhat.com>

Reviewed-by: Lukáš Slebodník <lslebodn@redhat.com>