INI: Prepare for schema validation
Pointer to function ini_schema_validator_func will be an optional
and can be used to schema validation. It shoul also prepare data
for validator function if last output argument is not NULL.
These prepared data will be passed to validator function as
a 5th argument of ini_validator_func. These two functions
are responsible for memory management of passed additional data.

It isn't an API/ABI change because lib_iniconfig has not been released yet.

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>