INI: Prevent null return_cfg during augment
This fixes the behavior of ini_config_augment so that result_cfg
will not be null unless out of memory. In particular, when base_cfg
is non-NULL and a fatal error occurred, result_cfg will now be a
copy of base_cfg. This reflects existing documentation, as base_cfg
will be augmented and result_cfg is the result of the merge.

Resolves:
https://pagure.io/SSSD/ding-libs/issue/2776

Reviewed-by: Michal Židek <mzidek@redhat.com>