INI: Add INI_MS_DETECT merge notifications
In ini_config_augment, INI_MS_DETECT is supposed to detect
duplicate sections. Previously this was exposed only as
a return status of EEXIST. This updates the behavior to
return an error_list with warnings of files containing
one or more duplicate sections.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>
Reviewed-by: Michal Židek <mzidek@redhat.com>

Merges: https://pagure.io/SSSD/ding-libs/issue/3167