INI: Allow longer values then PATH_MAX
The older API used fgets to read the values and the
size of buffer to hold the line was fixed. With the
new API we use getline to get the whole line and this
does not require pre-allocated buffer. The check
for lenght was not necessary here.

Resolves:
https://pagure.io/SSSD/ding-libs/issue/3177

Reviewed-by: Jakub Hrozek <jhrozek@redhat.com>

Merges: https://pagure.io/SSSD/ding-libs/pull-request/3178